refactor: general auth module refactoring
