+++ apps/api/src/app/auth/auth.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from "@nestjs/common"
 import { JwtService } from "@nestjs/jwt"
-import { AccountModel } from "../account/account.model"
-import { AccountService } from "../account/account.service"
-import { CreateAccountDTO } from "../account/dto/create-account.dto"
+import { AccountModel } from "../accounts/account.model"
+import { AccountService } from "../accounts/account.service"
+import { CreateAccountDTO } from "../accounts/dto/create-account.dto"
 
 @Injectable()
 export class AuthService {
@@ -10,22 +10,24 @@ export class AuthService {
         private readonly accountService: AccountService,
         private readonly jwtService: JwtService
     ) {}
+
     public async findOrCreateAccount(
         payload: CreateAccountDTO
     ): Promise<string> {
-        let user: AccountModel = await this.accountService.findOne({
+        let account: AccountModel = await this.accountService.findOne({
             username: payload.username,
             service: payload.service
         })
 
-        if (!user) {
-            user = await this.accountService.create(payload)
+        if (!account) {
+            account = await this.accountService.create(payload)
         }
-        return this.generateToken(user)
+
+        return this.generateToken(account)
     }
 
     public async generateToken(payload: AccountModel): Promise<string> {
-        return await this.jwtService.sign({
+        return this.jwtService.sign({
             id: payload.id,
             username: payload.username
         })
